Home screen UI issue resolved
